requirements.txt
boto3
fake_useragent
pyopenssl
requests
termcolor
pycryptodome